The current (incomplete) implementation of Extract_include_directives was simplified; the remaining member function to implement is the constructor.
The implementation of the stream extractor for  Extract_include_directives objects has been commented out,
since the work is now done by the constructor.
Regarding the \todo's inside this function, outlining the implementation: The thinking seems to be too "low level",
too much in the spirit of, not "Spirit", but in the spirit of
reading one character at a time. As we discussed it, grammars
must be written primarily with meaningful syntactical elements in mind.