C-corrections (enums, ODR, bool; parameters of main).
In C, the identifiers associated with an enum- or struct-type
are only tags, and must be accompanied with their specifications
when declaring variables.
And, again different from C++, constants defined at namespace scope
have (global) linkage, and thus can violate the ODR.
Finally, bool in C is only provided by the standard library.