Corrections for variables test_aux_dir, test_lib_dir, test_bin_dir.
"test-*" renamed to "test_*", and derived variables defined as
recursive variables.
Issue: "?=" cannot be used for variables depending on the module,
since (currently) these variables are defined twice when calling
the recursive makefile, and only the second time the value is right.

Also documentation related to these variables moved to the docus-file.