No automatic update of the relative-path-tool.
If it's not there, then it needs to be created, but otherwise
nothing happens --- updates must be handled manually.
The problem is, that the old dependency made the existence of
the prerequisite necessary, and there seems no (easy) possibility
to just express that if the prerequisite is there and newer, then
re-create, but otherwise just ignore it.