Providing ordered form of binary vdW-problems, and CORRECTION of polarity.
The positive clauses are associated with the first block, not with the
second (as before). In other words, all literals are flipped now (which
makes a change only for non-diagonal problems).
For solvers sensitive to polarities, for precise reproduction one needs
now to install an older version of the OKlibrary (prior to this change).