Re-opened todo "Fix translation system".
Ticket MG-22. Todos which have been marked as "DONE", and which have been
removed in this commit, must be checked whether they have been covered
completely. It doesn't look like that, but it looks as if a big blob,
"Fix translation system", was at some point just marked as "DONE", without
care for the details.