Request for MG to correct todo.
Ticket MG-78.
Speaking of the "latest" would only be correct, if that would be
constantly monitored. And likely such language should never be used,
when version-information is available, which then, of course, must be
given.
MG needs to remember that also todos are not verbal communications, but
are written documents, which must maintain their meaning over time.