Adding data on AES(r,2,1,8) using the forward MixColumns translation.
The forward (one directional) translation performs better than the bidirectional
for the canonical (~1.5x in terms of time and and 1.2x in terms of conflicts),
but worse for the 1-base translation (~2x worse for both time and conflicts).

The forward translation is slightly better when combined with the minimum
translation than the bidirectional, but not by much. These differences are
likely random effects.