Further updates to Experimentation milestones.
Given the scope of the "CLEANUP" milestone (0.4.2), the SAT2012 benchmarks should
be completed within this timeframe. The remaining tasks at
the Experimentation/Investigations level should be completed by the next
major milestone (0.4.4) of Experimentation.

The major milestone "UP TO DATE" is named such to signify that all tasks
involving updating old systems (e.g., OKgenerator) should be
synchonised with the rest of the library, and other modules should be in
an "up to date" state, with no untouched old todos.