Correction: correcting order of file data.
This was caused by using 1 rather 01 etc numbering in filenames.
This wasn't clear in the file as MG (wrongly) applied a substitution
on the 1-base file listing to get the min file listing.

All data should always be computed as it is stated, and assumptions of
"no-change" should *never* be made.