Removed restriction on clause-length.
Main step is removal of bitset-usage ("Clause_content"); its only purpose
was the determination of the unit-clause, and this can be done more
efficiently by just direct search.

Now the weights can be arbitrary real numbers. This incurred a (small)
speed loss, but that's worth it, since now the weights can be optimised.