Added lost todo comment.
The comment on reducing conditional checks was deleted without explanation in 801f7903ef93d884b69084b3c03ef52524abaaf3. It has been added to todos in a slightly improved form. The bug-report is updated regarding this issue and also regarding the whole style: in the recent commits all required comments were written in the form of complete sentences.