Fixed times for 0.19.1.
This is a fix of the issue found in cd9767e5a04a243828a74f6e27fde0a41151058f by OK on 30.10.2020. In particular, git-ids had become inconsistant for 0.19.1 on the server. The experiment was repeated and now everything should be ok. The results are very close to the previous ones.