Data update.
Results on 0.3.0 on amd1 were added. The summary on amd1 is almost the same as that on server: the columns x and min are the same, while the columns max, meand, and sd sloghtly differ. The results on 0.1.0 and 0.1.1 on amd1 were removed since it is more relevant to check results on the last version.