Basic lookahead propagator.
In this case a variable of the type IntVar should not be passed as an object, because each Gecode variable is connected with its initial Space. Using variable's array index i works, since here V[i] is a variable of the cloned space, so the constraint is added for the cloned space instead of the parent one.