Inner nodes counting fixed.
The number of inner nodes is the number of copies, but it is only when a search engine is initialises. Therefore one call of copy() spent for initialisation of the engine should not be counted. The calculated number of inner nodes corresponds to gist and to the Gecode Statistics. The number of solutions is also calculated correctly now. However, it seems that it is not possible to count the number of failed leaves (and, therefore, the total number of leaves) in the same way as it was done for the inner nodes. This counting happens inside the main function of the search engine.