Fixed general build script.
BRG is needed in app_tests of tawSolver, so now BRG is made before tawSolver in the general build scirpt. Also app_tests for DQCNF cause errors since some cnfs with unit clauses are run on tawSolver. Now for DQCNF app_tests are not made in the general build script to allow it finish the job with no errors.