Fixed number of inner nodes.
If a current node is already sat or unsat due to lookahead reduction, anyway according to Gecode it is treated like an inner node for which a leaf is created. This situation is tracked now and such an sat or unsat node is treated like a leaf in the backtracking tree. This fix does not affect the number of leaves.