Corrections (construction of ReductionStatistics, meaning of values.
Always look out for improving general facilities (obviously creating
a fake object m just to get a number into the statistics should not
be done).
We need to establish *meaning" for the classes in here.
"auto i = 0" is bad for many reasons ("auto" = int here, which should
be said, and which is the wrong type).